% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_methods.R
\name{forecast_CH_PeEn}
\alias{forecast_CH_PeEn}
\title{Do complete-history persistence ensemble forecast}
\usage{
forecast_CH_PeEn(GHI, percentiles, sun_up, clearsky_GHI, ts_per_hour)
}
\arguments{
\item{GHI}{A [day x time] matrix of the telemetry}

\item{percentiles}{A vector of the percentiles corresponding to the desired
forecast quantiles}

\item{sun_up}{A [day x time] matrix of logicals, indicating whether the sun
is up}

\item{clearsky_GHI}{a [day x time] matrix of clear-sky irradiance estimates}

\item{ts_per_hour}{Time-steps per hour, e.g., 12 for a 5-minute resolution
forecast}
}
\value{
a matrix of quantile forecasts at each valid time in the input data
}
\description{
Generates a CH-PeEn forecast using all clear-sky indices at the same hour of
day, for both intra-hourly and hourly-resolution forecasts. The clear-sky
index ensemble is multiplied by the forecast clear-sky GHI to get a GHI
ensemble, which is transformed to a full probabilistic forecast through an
empirical CDF.
}
\details{
Valid for both intra-hourly and hourly-resolution forecasts. This function is
for hindcasting only, and uses the same clear-sky GHI estimates for both the
historical and forecast values.
}
\seealso{
Other forecast functions: \code{\link{forecast_Gaussian_hourly}},
  \code{\link{forecast_Gaussian_intrahour}},
  \code{\link{forecast_NWP}},
  \code{\link{forecast_PeEn_hourly}},
  \code{\link{forecast_PeEn_intrahour}},
  \code{\link{forecast_climatology}}
}
\concept{forecast functions}
