% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_methods.R
\name{forecast_climatology}
\alias{forecast_climatology}
\title{Do climatology forecast}
\usage{
forecast_climatology(GHI, percentiles, sun_up)
}
\arguments{
\item{GHI}{A [day x time-step] matrix of the hourly average telemetry}

\item{percentiles}{A vector of the percentiles corresponding to the desired
forecast quantiles}

\item{sun_up}{A [day x time-step] matrix of logicals, indicating whether the
sun is up}
}
\value{
a matrix of quantile forecasts at each valid time in the input data
}
\description{
Generates a static climatology forecast using all GHI measurements in the
data set, transformed to a full probabilistic forecast using an empirical
CDF.
}
\details{
Valid for intra-hourly and hourly forecast resolutions
}
\seealso{
Other forecast functions: \code{\link{forecast_CH_PeEn}},
  \code{\link{forecast_Gaussian_hourly}},
  \code{\link{forecast_Gaussian_intrahour}},
  \code{\link{forecast_NWP}},
  \code{\link{forecast_PeEn_hourly}},
  \code{\link{forecast_PeEn_intrahour}}
}
\concept{forecast functions}
