% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_methods.R
\name{forecast_Gaussian_hourly}
\alias{forecast_Gaussian_hourly}
\title{Do hourly-resolution Gaussian error distribution forecast}
\usage{
forecast_Gaussian_hourly(nwp, GHI, percentiles, sun_up, clearsky_GHI)
}
\arguments{
\item{nwp}{A [day x issue time x lead time x member] matrix of NWP ensemble forecasts; First member is treated as the control member.}

\item{GHI}{A [day x time] matrix of the telemetry}

\item{percentiles}{A vector of the percentiles corresponding to the desired forecast quantiles}

\item{sun_up}{A [day x time] matrix of logicals, indicating whether the sun is up}

\item{clearsky_GHI}{a [day x time] matrix of clear-sky irradiance estimates}
}
\value{
a matrix of quantile forecasts at each valid time in the input data
}
\description{
Generates an hourly Gaussian error distribution forecast using a doubly
truncated Gaussian distribution. The distribution's mean is the deterministic
forecast from the control member (1st member) in the NWP ensemble. The
distribution's standard deviation is calculated as the standard deviation of
the deterministic forecast errors at the same hour-of-day over the sample
set.
}
\details{
Valid for hourly resolution forecasts. For intra-hourly forecasts, see
\code{\link{forecast_Gaussian_intrahour}}. Distribution is truncated at 0 on
the low end and clear-sky GHI on the upper end.
}
\seealso{
Other forecast functions: \code{\link{forecast_CH_PeEn}},
  \code{\link{forecast_Gaussian_intrahour}},
  \code{\link{forecast_NWP}},
  \code{\link{forecast_PeEn_hourly}},
  \code{\link{forecast_PeEn_intrahour}},
  \code{\link{forecast_climatology}},
  \code{\link{forecast_mcm}}
}
\concept{forecast functions}
