% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_methods.R
\name{forecast_mcm}
\alias{forecast_mcm}
\title{Do intra-hourly Markov-chain mixture distribution forecast}
\usage{
forecast_mcm(GHI, lead_up_GHI, percentiles, sun_up, clearsky_GHI,
  lead_up_clearsky_GHI, ts_per_hour, num_days,
  numBins = length(percentiles) + 1, numSamples = 1000,
  h_per_day = 24)
}
\arguments{
\item{GHI}{A vector of the telemetry}

\item{lead_up_GHI}{A vector of out-of-sample telemetry for the days leading
up to the start of the sample period. Number of days must be >= num_days}

\item{percentiles}{A vector of the percentiles corresponding to the desired
forecast quantiles}

\item{sun_up}{A vector of logicals, indicating whether the sun is up}

\item{clearsky_GHI}{a vector of clear-sky irradiance estimates}

\item{lead_up_clearsky_GHI}{A vector of out-of-sample clear-sky irradiance
estimates for the days leading up to the start of the sample period,
corresponding to lead_up_GHI}

\item{ts_per_hour}{Time-steps per hour, e.g., 12 for a 5-minute resolution
forecast}

\item{num_days}{Number of days of training data}

\item{numBins}{(optional) Number of bins to use in the MCM matrix M. Defaults
to number of percentiles + 1.}

\item{numSamples}{(optional) Number of samples to take from MCM model to
generate empirical CDF. Defaults to 1000.}

\item{h_per_day}{(optional) Hours per day = 24 (useful for testing)}
}
\description{
Valid for intra-hourly forecasts. This generates a clear-sky index transition
matrix based on the previous 20 days of data before each hourly issue time.
Using that transition matrix, clear-sky index is forecasted over the next D
time steps (i.e., up to 12 for a 5-minute resolution forecast.)
}
\details{
Modified from the main.R script at
https://github.com/SheperoMah/MCM-distribution-forecasting, Reported in: J.
Munkhammar, J. Widén, D. W. van der Meer, Probabilistic forecasting of
high-resolution clear-sky index time-series using a Markov-chain mixture
distribution model, Solar Energy vol. 184, pp. 688-695, 2019.
}
\seealso{
Other forecast functions: \code{\link{forecast_CH_PeEn}},
  \code{\link{forecast_Gaussian_hourly}},
  \code{\link{forecast_Gaussian_intrahour}},
  \code{\link{forecast_NWP}},
  \code{\link{forecast_PeEn_hourly}},
  \code{\link{forecast_PeEn_intrahour}},
  \code{\link{forecast_climatology}}
}
\concept{forecast functions}
