% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_methods.R
\name{forecast_NWP}
\alias{forecast_NWP}
\title{Do raw numerical weather prediction ensemble forecast}
\usage{
forecast_NWP(nwp, percentiles, sun_up)
}
\arguments{
\item{nwp}{A [day x issue time x lead time x member] matrix of NWP ensemble
forecasts}

\item{percentiles}{A vector of the percentiles corresponding to the desired
forecast quantiles}

\item{sun_up}{A [day x hour] matrix of logicals, indicating whether the sun
is up}
}
\value{
a matrix of quantile forecasts at each valid time in the input data
}
\description{
Generates a probabilistic forecast from a NwP ensemble by applying an
empirical CDF.
}
\details{
Valid at the resolution of the NWP ensemble (e.g., hourly)
}
\seealso{
Other forecast functions: \code{\link{forecast_CH_PeEn}},
  \code{\link{forecast_Gaussian_hourly}},
  \code{\link{forecast_Gaussian_intrahour}},
  \code{\link{forecast_PeEn_hourly}},
  \code{\link{forecast_PeEn_intrahour}},
  \code{\link{forecast_climatology}},
  \code{\link{forecast_mcm}}
}
\concept{forecast functions}
