% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_methods.R
\name{forecast_PeEn_hourly}
\alias{forecast_PeEn_hourly}
\title{Do hourly persistence ensemble forecast}
\usage{
forecast_PeEn_hourly(GHI, percentiles, sun_up, num_days, lead_up_GHI)
}
\arguments{
\item{GHI}{A [day x time] matrix of the hourly average telemetry}

\item{percentiles}{A vector of the percentiles corresponding to the desired
forecast quantiles}

\item{sun_up}{A [day x time] matrix of logicals, indicating whether the sun
is up}

\item{num_days}{Number of days of data to take, indicating size of
persistence ensemble}

\item{lead_up_GHI}{A [day x time] matrix of out-of-sample telemetry for the
days leading up to the start of the sample period. Number of days must be
>= num_days}
}
\value{
a matrix of quantile forecasts at each valid time in the input data
}
\description{
Generates an hourly persistence ensemble forecast, using GHI measurements
from the same hour-of-day over the last few days. The GHI ensemble is
transformed to a full probabilistic forecast using an empirical CDF.
}
\details{
Valid for hourly resolution forecasts. For intra-hourly forecasts, see
\code{\link{forecast_PeEn_intrahour}}.
}
\seealso{
Other forecast functions: \code{\link{forecast_CH_PeEn}},
  \code{\link{forecast_Gaussian_hourly}},
  \code{\link{forecast_Gaussian_intrahour}},
  \code{\link{forecast_NWP}},
  \code{\link{forecast_PeEn_intrahour}},
  \code{\link{forecast_climatology}}
}
\concept{forecast functions}
