% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{interval_width}
\alias{interval_width}
\title{Calculate average interval width}
\usage{
interval_width(fc, sun_up, intervals = seq(0.1, 0.9, by = 0.1))
}
\arguments{
\item{fc}{A [valid time x quantile] matrix of probabilistic quantile
forecasts, with column names giving the [0,1] percentiles of the forecast}

\item{sun_up}{Logical vector of whether sun is up across the forecast times}

\item{intervals}{A vector of central intervals in (0,1) whose widths should
be calculated. Should correspond to available percentile columns in fc.}
}
\value{
a list of the intervals and their average widths
}
\description{
Calculate average sharpness for central intervals from 10% to 90%. Negatively
oriented (smaller is better).
}
