% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plot_fanplot}
\alias{plot_fanplot}
\title{Plot fan plot of time-series forecasts}
\usage{
plot_fanplot(fc, tel, window, ts_per_hour, output_directory, site, res,
  method, R_graph_export, ymax = 1200)
}
\arguments{
\item{fc}{A [valid time x quantile] matrix of probabilistic quantile
forecasts, with column names giving the [0,1] percentiles of the forecast}

\item{tel}{A vector of the telemetry values}

\item{window}{A (min, max) vector of time indices, corresponding to the rows
of fc to plot}

\item{ts_per_hour}{Time-steps per hour}

\item{output_directory}{Directory to save graphs}

\item{site}{Site name, for naming output file}

\item{res}{Temporal resolution, for naming output file}

\item{method}{Forecast method name, for naming output file}

\item{R_graph_export}{Boolean, whether to save a .R object of the plot as
well}

\item{ymax}{Y-axis maximum value (defaults to 1200 W/m^2)}
}
\description{
Plot a fan plot comparing the forecast quantiles to the telemetry
}
