% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plot_interval_width}
\alias{plot_interval_width}
\title{Plot average interval width}
\usage{
plot_interval_width(interval_width_df, site, res, shapes, output_directory,
  R_graph_export)
}
\arguments{
\item{interval_width_df}{Data frame of the interval width results for each
method}

\item{site}{Site name, for naming output file}

\item{res}{Temporal resolution, for naming output file}

\item{shapes}{List of ggplot2 shapes, one for each method}

\item{output_directory}{Directory to save graphs}

\item{R_graph_export}{Boolean, whether to save a .R object of the plot as
well}
}
\description{
Plot average interval width of 10%, ..., 90% central intervals, comparing
among forecast methods for a single site
}
