% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plot_PIT_histogram}
\alias{plot_PIT_histogram}
\title{Plot probability integral transform histogram}
\usage{
plot_PIT_histogram(fc, tel, sun_up, nbins, site, res, method,
  output_directory, R_graph_export)
}
\arguments{
\item{fc}{A [valid time x quantile] matrix of probabilistic quantile
forecasts, with column names giving the [0,1] percentiles of the forecast}

\item{tel}{A vector of the telemetry values}

\item{sun_up}{Logical vector of whether sun is up across the forecast times}

\item{nbins}{The number of histogram bins to use. Must be a factor of
length(percentiles)+1 to make use of available fc format.}

\item{site}{Site name, for naming output file}

\item{res}{Temporal resolution, for naming output file}

\item{method}{Forecast method name, for naming output file}

\item{output_directory}{Directory to save graphs}

\item{R_graph_export}{Boolean, whether to save a .R object of the plot as
well}
}
\description{
Plot PIT histogram for a single single and forecast method
}
