% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plot_quantile_score}
\alias{plot_quantile_score}
\title{Plot quantile score vs. quantile}
\usage{
plot_quantile_score(fc, tel, sun_up, percentiles, site, res, method,
  output_directory, R_graph_export)
}
\arguments{
\item{fc}{A [valid time x quantile] matrix of probabilistic quantile
forecasts, with column names giving the [0,1] percentiles of the forecast}

\item{tel}{A vector of the telemetry values}

\item{sun_up}{Logical vector of whether sun is up across the forecast times}

\item{percentiles}{A vector (0,1) corresponding to the columns of fc}

\item{site}{Site name, for naming output file}

\item{res}{Temporal resolution, for naming output file}

\item{method}{Forecast method name, for naming output file}

\item{output_directory}{Directory to save graphs}

\item{R_graph_export}{Boolean, whether to save a .R object of the plot as
well}
}
\description{
Plot unweighted, left-, and right-weighted quantile scores of selected
quantiles
}
