% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{plot_reliability}
\alias{plot_reliability}
\title{Plot average quantile reliability}
\usage{
plot_reliability(reliability_df, site, res, shapes, output_directory,
  R_graph_export)
}
\arguments{
\item{reliability_df}{Data frame of the quantile reliability results for each
method}

\item{site}{Site name, for naming output file}

\item{res}{Temporal resolution, for naming output file}

\item{shapes}{List of ggplot2 shapes, one for each method}

\item{output_directory}{Directory to save graphs}

\item{R_graph_export}{Boolean, whether to save a .R object of the plot as
well}
}
\description{
Plot reliability of 1..99 quantiles, comparing among forecast methods for a
single site
}
