% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{QS}
\alias{QS}
\title{Get quantile score at each quantile}
\usage{
QS(fc, tel, sun_up, percentiles)
}
\arguments{
\item{fc}{A [valid time x quantile] matrix of probabilistic quantile forecasts}

\item{tel}{A vector of the telemetry values}

\item{sun_up}{Logical vector of whether sun is up across the forecast times}

\item{percentiles}{A vector (0,1) corresponding to the columns of fc}
}
\value{
the quantile scores at each quantile in the fc matrix
}
\description{
Calculate quantile score at each quantile in the fc matrix
}
