% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{qwCRPS}
\alias{qwCRPS}
\title{Calculate quantile-weighted CRPS}
\usage{
qwCRPS(fc, tel, sun_up, weighting = "none")
}
\arguments{
\item{fc}{A [valid time x quantile] matrix of probabilistic quantile
forecasts, with column names giving the [0,1] percentiles of the forecast}

\item{tel}{A vector of the telemetry values}

\item{sun_up}{Logical vector of whether sun is up across the forecast times}

\item{weighting}{One of "none" (default), "tails", "right", "left", "center"}
}
\value{
The weighted CRPS
}
\description{
Get the quantile-weighted CRPS (continuous ranked probability score) for the
forecast, estimated by trapezoidal integration of its quantile decomposition.
Can be weighted or unweighted (default).
}
