% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{reliability}
\alias{reliability}
\title{Calculate reliability at each quantile}
\usage{
reliability(fc, tel, sun_up)
}
\arguments{
\item{fc}{A [valid time x quantile] matrix of probabilistic quantile
forecasts, with column names giving the [0,1] percentiles of the forecast}

\item{tel}{A vector of the telemetry values}

\item{sun_up}{Logical vector of whether sun is up across the forecast times}
}
\value{
A vector of observed proportions corresponding to each column in fc
}
\description{
Calculate proportion of observations that fell within a given quantile,
relative to each nominal proportion (percentile).
}
